Fix potential stream left open
Previous to this commit, it was possible for open input streams,
processes, and bufferred streams to remain open if a failure happened
immediately after the streams and processes was created. This commit
adds a `finally` block to the `try` statement that ensures they are
closed if they're open.

This commit also fixes an issue with concatonating strings in a loop.