Make PuppetEnterpriseConfig a static class
Previous to this commit, the PuppetEnterpriseStep class created a
PuppetEnterpriseConfig class object to retrieve the PE configs (master
address and CA certificate). This made testing very difficult since it's
best to set the values for the CA certificate and master address in the
test and then call the necessary step methods in the test.

This commit makes the PuppetEtnerpriseConfig class contain all static
methods so any objects interacting with the class will be sharing the
same class data.