Mock Puppet CA API + code cleanup
Previous to this commit, the CA certificate used for communicating with
the mocked Puppet Orchestrator API service was hard coded. This meant
that we were not validating that the plugin could succesfully retrieve
the CA certificate from the Puppet Server's CA API and configure the
plugin to use it.  This commit mocks the Puppet Server CA API listening
on port 8140 and responds with the correct CA certificate used to serve
the mocked Puppet Orchestrator API.  This validates end-to-end
communication with the necessary Puppet services.

This commit also moves large strings such as the CA certificate string
and long API call responses to files. Some uncommented code was also
removed since it wasn't necessary anymore