Use Google's Gson instead of quick-json
Previous to this commit, quick-json was used to parse JSON from Puppet
Enterprise API calls. However, it turns out that library breaks on empty
arrays. It always assumes there should be a value. This commit replaces
the quick-json library with Google's Gson library.

One gotcha was Gson uses LinkedTreeMap for JSON hashes instead of
HashMap, so each instance of HashMap needed to be replaced.