Refactor PERequest class
This commit refactors the PERequest class to be an abstract class that
other classes can extend. That way different objects meant to interact
with the PE APIs can all use a common transport to work from,
dramatically simplifying the model.

This commit also renames the `request` method to `send` and generally
cleans things up.