Allow users to specify empty nodes list
This commit enables users to give an empty nodes array for the 'nodes'
parameter in the puppet.job step. Previously, if the list was empty, the
parameter would not have been passed to the API call that creates the
job. Since there would be no targets, the Puppet orchestrator would
target the entire environment, not zero nodes which is what would be
expected.